[ROCm][MiniMax-M3] Support an fp8 indexer KV cache
The AMD MiniMax-M3 path forced a bf16 indexer side cache. Make the Triton
index kernels dtype-agnostic (widen the cache to the query dtype before
tl.dot, cast explicitly on the side-cache store) and plumb
attention_config.indexer_kv_dtype through MiniMaxM3SparseAttention, so
--attention-config '{"indexer_kv_dtype":"fp8"}' works on ROCm.

fp8 halves the index-cache read, which dominates the indexer's decode cost
at high concurrency. MiniMax-M3-MXFP4, MI355X, TP4, 8k/1k: +2.9% at c64
(22,401 -> 23,058 tok/s) and +10.6% at c128. GSM8K 5-shot strict 0.9416
(n=1319, stderr ~0.006).

Requires VLLM_ROCM_SHUFFLE_KV_CACHE_LAYOUT=1: only the AITER sparse-PA path
writes the side cache through the casting Triton insert, while
fused_minimax_m3_qknorm_rope_kv_insert stores bf16 unconditionally. The
layer raises if that combination is requested rather than silently keeping
bf16 keys.

+++ vllm/models/minimax_m3/amd/ops/index_topk.py
@@ -146,7 +146,7 @@ def _index_block_score_kernel(
             + off_k[None, :] * stride_ik_pos
             + off_d[:, None] * stride_ik_d,
         )
-        qk = tl.dot(q, k)
+        qk = tl.dot(q, k.to(q.dtype), out_dtype=tl.float32)
         # apply causal mask as needed
         if q_start < i + BLOCK_SIZE_K:
             qk = tl.where(off_q[:, None] >= pos[None, :], qk, float("-inf"))
@@ -383,7 +383,7 @@ def _decode_index_score_kernel(
             # are loaded in their stored dtype (bf16 or e4m3) and the MMA
             # accumulates in fp32 so the per-block max score is exact for the
             # fp8 indexer too.
-            kq = tl.dot(k, q, out_dtype=tl.float32)  # [N,HQ]
+            kq = tl.dot(k.to(q.dtype), q, out_dtype=tl.float32)  # [N,HQ]
         kq = tl.where(pos_mask & q_mask[None, :], kq, float("-inf"))
         score = tl.max(kq, axis=0)  # [HQ]
         is_visible_block = blk < num_blocks_q
